<fix>[zstackctl]: upgrade python3
upgrade zstack-ctl zstack-cli
kvmagent cephagent consoleproxy
to python 3.11

upgrade most of dependency

fix most of warning.

Resolves: ZSTAC-73161

+++ kvmagent/kvmagent/test/shareblock_testsuite/test_sharedblock_convert_volume_provisioning.py
@@ -53,12 +53,12 @@ class TestSharedBlockPlugin(TestCase, SharedBlockPluginTestStub):
         )
         self.assertEqual(True, rsp.success, rsp.error)
 
-        old_size = long(lvm.get_lv_size("/dev/%s/%s" % (vgUuid, volume_uuid)))
+        old_size = int(lvm.get_lv_size("/dev/%s/%s" % (vgUuid, volume_uuid)))
         rsp = sharedblock_utils.sharedblock_convert_volume_provisioning(vgUuid, hostUuid, volumeUuid=volume_uuid,
                                                                         installPath="sharedblock://{}/{}".format(vgUuid, volume_uuid),
                                                                         provisioningStrategy='ThinProvisioning',
                                                                         addons={"thinProvisioningInitializeSize":1048576*1024*5})
 
         self.assertEqual(True, rsp.success, rsp.error)
-        new_size = long(lvm.get_lv_size("/dev/%s/%s" % (vgUuid, volume_uuid)))
+        new_size = int(lvm.get_lv_size("/dev/%s/%s" % (vgUuid, volume_uuid)))
         self.assertLess(new_size, old_size)
